backend/pyproject.toml
[project]
name = "treasure"
version = "0.1.0"
description = "Personal Finance Tracker — backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=23.0.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "cryptography>=43.0.0",
    "beanie>=1.27.0",
    "fastapi-users[beanie]>=14.0.0",
    "pwdlib[argon2]>=0.2.0",
    "python-multipart>=0.0.12",
]

[project.scripts]
treasure-setup = "app.cli.setup:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["app*"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
# B008: Depends() in default args is the standard FastAPI pattern — not a bug.
ignore = ["B008"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
